[Decision Trees] Added Huber loss function implementation
Summary:
A fairly vanilla implementation from http://www-stat.stanford.edu/~jhf/ftp/trebst.pdf.

There's a fair amount of jankiness due to the need to pass global information around
to the loss function. An alternative may be to pass the examples directly in the constructor
of the loss function, and recompute deltaM in that form.

Test Plan: no unit tests yet

Differential Revision: http://phabricator.tullo.ch/D8